requirements.txt
streamlit
pillow
google-generativeai
gtts